exchange: delay_sync_on_auth_error
If an auth error happens while last ping is active and server request a new sync,
next syncs will enter in a loop, asking the server for a sync operation every minute.
In that case increase the timeout to don't kill the battery. A notification will delivery to
user so he was inform of the auth error (and in anycase we reduce the sync operations
to one every 30 minutes until auth error is resolved).

Change-Id: I91bafb08ba6e366c558ee37b2afb890b88cd4315
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>